refactor: move auth to fixture folder
